+++ demo/taxonomy-browser/taxonomy-browser-itis.js
@@ -1,7 +1,7 @@
 /*!
  * Fancytree Taxonomy Browser
  *
- * Copyright (c) 2015, Martin Wendt (http://wwWendt.de)
+ * Copyright (c) 2015, Martin Wendt (https://wwWendt.de)
  *
  * Released under the MIT license
  * https://github.com/mar10/fancytree/wiki/LicenseInfo