+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-# 2.30.3-0 / Unreleased
+# 2.31.0-0 / Unreleased
+  * [Added] New tree option `.treeId`
+  * [Changed] `.getTree()` now also accepts the tree id string
   * Replace jshint/jscs with eslint
   * Now testing on Puppeteer/Chromium instead of PhantonJS
-  * Deprecated loaderror and lazyload options now throw an error
+  * [DEPRECATED] loaderror and lazyload options now throw an error
+
 
 # 2.30.2 / 2019-01-13
   * Stop testing on IE 8 (no longer available on Saucelabs)